INS-2531: moved test functions to test files
Functions that are only used for testing are now moved to _test.go files
